experimental fix to issue #10 and wuub's issue #204
The old clojure_sender code wraps the evaluated expression in a load-string
call, which does not update *ns*. Presumably this needed to be done for the old repl where every line in the evaluated region would be sent separately, but this is not the case with nrepl. Additionally, the wrapper introduced a double escaping problem in wuub's issue #204.